separate StashDate
