+++ src/content.js
@@ -3,6 +3,7 @@
 import React, { useState, useEffect } from "react";
 import ReactDOM from "react-dom";
 import ReactModal from "react-modal";
+import StashDate from "./components/StashDate";
 import StashEntries from "./components/StashEntries";
 import "./css/content.css";
 import {
@@ -193,57 +194,6 @@ function StashList({data, chromePort}) {
 
 function Separator() {
   return <div class="border-shade-87 border-top-normal"></div>
-}
-
-const daysOfWeek = [
-  "Monday",
-  "Tuesday",
-  "Wednesday",
-  "Thursday",
-  "Friday",
-  "Saturday",
-  "Sunday"
-];
-
-const months = [
-  "January",
-  "February",
-  "March",
-  "April",
-  "May",
-  "June",
-  "July",
-  "August",
-  "September",
-  "October",
-  "November",
-  "December"
-];
-
-function StashDate({fullYear, month, date, day}) {
-  function displayDate(fullYear, month, date, day) {
-    const today = new Date();
-    const thatDate = {fullYear, month, date}
-    const commonPart = `${daysOfWeek[day]}, ${months[month]} ${date}`;
-    if (hasSameDate(thatDate, today))
-      return `Today - ${commonPart}`;
-    today.setDate(today.getDate() - 1);
-    if (hasSameDate(thatDate, today))
-      return `Yesterday - ${commonPart}`;
-    if (fullYear == today.getFullYear()) return `${commonPart}`;
-    else return `${commonPart}, ${fullYear}`;
-  }
-  return (
-    <div className={"flexbox flexbox-justify-center"}>
-      <div
-        className={
-          "shade-087 font-size-huge font-weight-bold line-height-huge padding-top-extra-larger"
-        }
-      >
-        {displayDate(fullYear, month, date, day)}
-      </div>
-    </div>
-  );
 }
 
 const app = document.createElement("div");
